Show the records already in the collection (#31, #41)
This adds a client-side router. For now, it only serves to check
whether existing records should be shown, and if so, causes them to be
shown. However, it should be fairly straightforward to add more
routes.

Overall routing is still concentrated at the server side. As long as
this is the case, state management is bound to be a bit kludgy.