Decouple and sanitize frontend RecordDetailView (#123 #130 #131)
Instead of reusing the same global view for subsequent presentations,
this mediates the creation and destruction of short-lived local views
through a radio channel and an aspect module. As this reduced the
burden of state management considerably, I took the opportunity to
self-render the view and close existing memory leaks. As a bonus, the
behavior seems more reliable now.